# openapi.yaml
openapi: 3.0.3
info:
  title: DevStandup.ai API
  description: AI-powered engineering team assistant for standups, code reviews, and documentation
  version: 1.0.0
  contact:
    name: DevStandup.ai Support
    url: https://github.com/devstandup-ai/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.devstandup.ai
    description: Production server

paths:
  /health:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/standup:
    get:
      summary: Get team standup summary
      operationId: getStandup
      tags:
        - Standup
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Standup summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamStandup'
        '404':
          description: No standup found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/standup/generate:
    post:
      summary: Generate new standup summary
      operationId: generateStandup
      tags:
        - Standup
      parameters:
        - name: since_hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Standup generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  standup:
                    $ref: '#/components/schemas/TeamStandup'

  /api/reviews:
    get:
      summary: Get code reviews
      operationId: getReviews
      tags:
        - Reviews
      parameters:
        - name: pr_number
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CodeReview'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CodeReview'

  /api/reviews/generate:
    post:
      summary: Generate code review
      operationId: generateReview
      tags:
        - Reviews
      parameters:
        - name: pr_number
          in: query
          schema:
            type: integer
        - name: since_hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Review generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/CodeReview'

  /api/docs/generate:
    post:
      summary: Generate documentation
      operationId: generateDocs
      tags:
        - Documentation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Documentation generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  docstring:
                    $ref: '#/components/schemas/DocstringGeneration'

  /api/tests/generate:
    post:
      summary: Generate unit tests
      operationId: generateTests
      tags:
        - Testing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Tests generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  test:
                    $ref: '#/components/schemas/TestGeneration'

  /api/activity:
    get:
      summary: Get developer activity
      operationId: getActivity
      tags:
        - Activity
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Activity retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeveloperActivity'
                  count:
                    type: integer

  /api/activity/aggregate:
    post:
      summary: Aggregate activity from GitHub
      operationId: aggregateActivity
      tags:
        - Activity
      parameters:
        - name: since_hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Activity aggregated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  activities_count:
                    type: integer

components:
  schemas:
    TeamStandup:
      type: object
      properties:
        date:
          type: string
          format: date-time
        team_items:
          type: array
          items:
            $ref: '#/components/schemas/StandupItem'
        summary:
          type: string
        key_highlights:
          type: array
          items:
            type: string
        team_blockers:
          type: array
          items:
            type: string
        total_developers:
          type: integer
        developers_with_blockers:
          type: integer

    StandupItem:
      type: object
      properties:
        developer:
          type: string
        yesterday:
          type: array
          items:
            type: string
        today:
          type: array
          items:
            type: string
        blockers:
          type: array
          items:
            type: string
        mood:
          type: string

    CodeReview:
      type: object
      properties:
        pr_number:
          type: integer
        pr_title:
          type: string
        reviewer:
          type: string
        timestamp:
          type: string
          format: date-time
        summary:
          type: string
        overall_score:
          type: integer
          minimum: 1
          maximum: 10
        findings:
          type: array
          items:
            $ref: '#/components/schemas/ReviewFinding'
        suggestions:
          type: array
          items:
            type: string
        approved:
          type: boolean
        critical_issues:
          type: integer
        high_issues:
          type: integer

    ReviewFinding:
      type: object
      properties:
        file_path:
          type: string
        line_number:
          type: integer
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        category:
          type: string
        message:
          type: string
        suggestion:
          type: string
        code_snippet:
          type: string

    DeveloperActivity:
      type: object
      properties:
        developer:
          type: string
        date:
          type: string
          format: date-time
        total_commits:
          type: integer
        total_prs:
          type: integer
        total_issues:
          type: integer
        commits:
          type: array
          items:
            $ref: '#/components/schemas/GitHubCommit'
        pull_requests:
          type: array
          items:
            $ref: '#/components/schemas/GitHubPullRequest'
        issues:
          type: array
          items:
            $ref: '#/components/schemas/GitHubIssue'

    GitHubCommit:
      type: object
      properties:
        sha:
          type: string
        message:
          type: string
        author:
          type: string
        timestamp:
          type: string
          format: date-time
        url:
          type: string
        files_changed:
          type: array
          items:
            type: string

    GitHubPullRequest:
      type: object
      properties:
        number:
          type: integer
        title:
          type: string
        body:
          type: string
        author:
          type: string
        state:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        url:
          type: string
        diff_url:
          type: string
        files_changed:
          type: array
          items:
            type: string
        additions:
          type: integer
        deletions:
          type: integer

    GitHubIssue:
      type: object
      properties:
        number:
          type: integer
        title:
          type: string
        body:
          type: string
        author:
          type: string
        state:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        url:
          type: string
        labels:
          type: array
          items:
            type: string

    DocstringGeneration:
      type: object
      properties:
        file_path:
          type: string
        function_name:
          type: string
        original_code:
          type: string
        generated_docstring:
          type: string
        style:
          type: string

    TestGeneration:
      type: object
      properties:
        file_path:
          type: string
        function_name:
          type: string
        original_code:
          type: string
        generated_test:
          type: string
        test_framework:
          type: string
        coverage_estimate:
          type: integer

    GenerateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [standup, review, docstring, test]
        data:
          type: object
      required:
        - type
        - data

    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string